pyproject.toml
[tool.poetry]
name = "todo-server"
version = "0.1.0"
description = ""
authors = ["Natthaphon Phanit <4550615-kekhuay@users.noreply.gitlab.com>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.27.2"
grpcio-tools = "^1.27.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
